requirements.txt
asgiref==3.7.2
Django==4.1.11
django-cors-headers==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
mongoengine==0.27.0
pymongo==3.12.1
pytz==2023.3.post1
sqlparse==0.2.4
typing_extensions==4.7.1



#pip freeze > requirements.txt